Remove SimpleClient, Producer, Consumer, Unittest (#1196)
In the 2.0 release, we're removing:
 * `SimpleClient`
 * `SimpleConsumer`
 * `SimpleProducer`
 * Old partitioners used by `SimpleProducer`; these are superceded by
 the `DefaultPartitioner`

These have been deprecated for several years in favor of `KafkaClient`
/ `KafkaConsumer` / `KafkaProducer`.

Since 2.0 allows breaking changes, we are removing the deprecated
classes.

Additionally, since the only usage of `unittest` was in tests for these
old Simple* clients, this also drops `unittest` from the library. All
tests now run under `pytest`.
