--- docs/apidoc/kafka.consumer.rst
@@ -1,46 +1,0 @@
-kafka.consumer package
-======================
-
-Submodules
-----------
-
-kafka.consumer.base module
---------------------------
-
-.. automodule:: kafka.consumer.base
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.consumer.kafka module
----------------------------
-
-.. automodule:: kafka.consumer.kafka
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.consumer.multiprocess module
-----------------------------------
-
-.. automodule:: kafka.consumer.multiprocess
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.consumer.simple module
-----------------------------
-
-.. automodule:: kafka.consumer.simple
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: kafka.consumer
-    :members:
-    :undoc-members:
-    :show-inheritance:
