--- docs/apidoc/kafka.coordinator.assignors.rst
@@ -1,30 +1,0 @@
-kafka.coordinator.assignors package
-===================================
-
-Submodules
-----------
-
-kafka.coordinator.assignors.abstract module
--------------------------------------------
-
-.. automodule:: kafka.coordinator.assignors.abstract
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.coordinator.assignors.roundrobin module
----------------------------------------------
-
-.. automodule:: kafka.coordinator.assignors.roundrobin
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: kafka.coordinator.assignors
-    :members:
-    :undoc-members:
-    :show-inheritance:
