--- docs/apidoc/kafka.coordinator.rst
@@ -1,45 +1,0 @@
-kafka.coordinator package
-=========================
-
-Subpackages
------------
-
-.. toctree::
-
-    kafka.coordinator.assignors
-
-Submodules
-----------
-
-kafka.coordinator.base module
------------------------------
-
-.. automodule:: kafka.coordinator.base
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.coordinator.consumer module
----------------------------------
-
-.. automodule:: kafka.coordinator.consumer
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.coordinator.heartbeat module
-----------------------------------
-
-.. automodule:: kafka.coordinator.heartbeat
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: kafka.coordinator
-    :members:
-    :undoc-members:
-    :show-inheritance:
