--- docs/apidoc/kafka.partitioner.rst
@@ -1,38 +1,0 @@
-kafka.partitioner package
-=========================
-
-Submodules
-----------
-
-kafka.partitioner.base module
------------------------------
-
-.. automodule:: kafka.partitioner.base
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.partitioner.hashed module
--------------------------------
-
-.. automodule:: kafka.partitioner.hashed
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.partitioner.roundrobin module
------------------------------------
-
-.. automodule:: kafka.partitioner.roundrobin
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: kafka.partitioner
-    :members:
-    :undoc-members:
-    :show-inheritance:
