--- docs/apidoc/kafka.producer.rst
@@ -1,38 +1,0 @@
-kafka.producer package
-======================
-
-Submodules
-----------
-
-kafka.producer.base module
---------------------------
-
-.. automodule:: kafka.producer.base
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.producer.keyed module
----------------------------
-
-.. automodule:: kafka.producer.keyed
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.producer.simple module
-----------------------------
-
-.. automodule:: kafka.producer.simple
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: kafka.producer
-    :members:
-    :undoc-members:
-    :show-inheritance:
