--- docs/apidoc/kafka.protocol.rst
@@ -1,126 +1,0 @@
-kafka.protocol package
-======================
-
-Submodules
-----------
-
-kafka.protocol.abstract module
-------------------------------
-
-.. automodule:: kafka.protocol.abstract
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.protocol.admin module
----------------------------
-
-.. automodule:: kafka.protocol.admin
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.protocol.api module
--------------------------
-
-.. automodule:: kafka.protocol.api
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.protocol.commit module
-----------------------------
-
-.. automodule:: kafka.protocol.commit
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.protocol.fetch module
----------------------------
-
-.. automodule:: kafka.protocol.fetch
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.protocol.group module
----------------------------
-
-.. automodule:: kafka.protocol.group
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.protocol.legacy module
-----------------------------
-
-.. automodule:: kafka.protocol.legacy
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.protocol.message module
------------------------------
-
-.. automodule:: kafka.protocol.message
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.protocol.metadata module
-------------------------------
-
-.. automodule:: kafka.protocol.metadata
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.protocol.offset module
-----------------------------
-
-.. automodule:: kafka.protocol.offset
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.protocol.pickle module
-----------------------------
-
-.. automodule:: kafka.protocol.pickle
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.protocol.produce module
------------------------------
-
-.. automodule:: kafka.protocol.produce
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.protocol.struct module
-----------------------------
-
-.. automodule:: kafka.protocol.struct
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.protocol.types module
----------------------------
-
-.. automodule:: kafka.protocol.types
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: kafka.protocol
-    :members:
-    :undoc-members:
-    :show-inheritance:
