--- docs/apidoc/kafka.rst
@@ -1,89 +1,0 @@
-kafka package
-=============
-
-Subpackages
------------
-
-.. toctree::
-
-    kafka.cluster
-    kafka.consumer
-    kafka.partitioner
-    kafka.producer
-
-Submodules
-----------
-
-kafka.cluster module
---------------------
-
-.. automodule:: kafka.cluster
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-kafka.client module
--------------------
-
-.. automodule:: kafka.client
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.codec module
-------------------
-
-.. automodule:: kafka.codec
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.common module
--------------------
-
-.. automodule:: kafka.common
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.conn module
------------------
-
-.. automodule:: kafka.conn
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.context module
---------------------
-
-.. automodule:: kafka.context
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.protocol module
----------------------
-
-.. automodule:: kafka.protocol
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-kafka.util module
------------------
-
-.. automodule:: kafka.util
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: kafka
-    :members:
-    :undoc-members:
-    :show-inheritance:
