--- kafka/common.py
@@ -1,4 +1,0 @@
-from __future__ import absolute_import
-
-from kafka.structs import *
-from kafka.errors import *
