+++ kafka/consumer/__init__.py
@@ -1,9 +1,7 @@
 from __future__ import absolute_import
 
-from kafka.consumer.simple import SimpleConsumer
-from kafka.consumer.multiprocess import MultiProcessConsumer
 from kafka.consumer.group import KafkaConsumer
 
 __all__ = [
-    'SimpleConsumer', 'MultiProcessConsumer', 'KafkaConsumer'
+    'KafkaConsumer'
 ]
