+++ kafka/producer/__init__.py
@@ -1,10 +1,7 @@
 from __future__ import absolute_import
 
 from kafka.producer.kafka import KafkaProducer
-from kafka.producer.simple import SimpleProducer
-from kafka.producer.keyed import KeyedProducer
 
 __all__ = [
-    'KafkaProducer',
-    'SimpleProducer', 'KeyedProducer' # deprecated
+    'KafkaProducer'
 ]
