+++ tox.ini
@@ -3,7 +3,6 @@ envlist = py{26,27,34,35,36,37,py}, docs
 
 [pytest]
 testpaths = kafka test
-doctest_optionflags = modules
 addopts = --durations=10
 log_format = %(created)f %(filename)-23s %(threadName)s %(message)s
 
@@ -19,7 +18,6 @@ deps =
     lz4
     xxhash
     crc32c
-    py26: unittest2
 commands =
     py.test {posargs:--pylint --pylint-rcfile=pylint.rc --pylint-error-types=EF --cov=kafka --cov-config=.covrc}
 setenv =
